requirements.txt
ansiwrap==0.8.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
console-menu==0.7.0
ConsoleMenu==1.0.1
getfilelistpy==1.0.7
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
items==0.6.5
nulltype==2.3.1
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.7
python-magic==0.4.25
python-magic-bin==0.4.14
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.8
simple-term-menu==1.4.1
six==1.16.0
textwrap3==0.9.2
uritemplate==4.1.1
urllib3==1.26.9